add pounder example
