+++ examples/driver/driver_example.go
@@ -21,9 +21,10 @@ package main
 import "database/sql"
 import (
 	"fmt"
-	_ "github.com/VoltDB/voltdb-client-go/voltdbclient"
 	"log"
 	"os"
+
+	_ "github.com/VoltDB/voltdb-client-go/voltdbclient"
 )
 
 func main() {
@@ -40,21 +41,6 @@ func main() {
 		log.Fatal(err)
 		os.Exit(-1)
 	}
-
-	stmt, err := db.Prepare("{call Vote(?, ?, ?)}")
-	if err != nil {
-		log.Fatal(err)
-		os.Exit(-1)
-	}
-
-	rows, err := stmt.Query( 1,2,3)
-	if err != nil {
-		log.Fatal(err)
-		os.Exit(-1)
-	}
-	printRows(rows)
-
-	/*
 	rows, err := db.Query("HELLOWORLD.select", "French")
 	if err != nil {
 		log.Fatal(err)
@@ -69,24 +55,25 @@ func main() {
 		os.Exit(-1)
 	}
 
-	rows, err = stmt.Query("c")
+	rows, err = stmt.Query("French")
 	if err != nil {
 		log.Fatal(err)
 		os.Exit(-1)
 	}
 	printRows(rows)
-	*/
 
 }
 
 func printRows(rows *sql.Rows) {
 	for rows.Next() {
-		var hello  int64
-		err := rows.Scan(&hello)
+		var hello string
+		var world string
+		var dialect string
+		err := rows.Scan(&hello, &world, &dialect)
 		if err != nil {
 			fmt.Println(err)
 			break
 		}
-		fmt.Printf("SUCCESS: %s %s %s\n", hello)
+		fmt.Printf("SUCCESS: %s %s %s\n", hello, world, dialect)
 	}
 }
