+++ examples/voter/.gitignore
@@ -1,0 +1,2 @@
+*.class
+*.log
