+++ voltdbclient/rows.go
@@ -171,7 +171,7 @@ func (vr VoltRows) AdvanceToRow(rowIndex int32) bool {
 
 // Advances to the next table.  Returns false if there isn't a next table.
 func (vr VoltRows) AdvanceTable() bool {
-	return vr.AdvanceToTable(vr.tableIndex+1)
+	return vr.AdvanceToTable(vr.tableIndex + 1)
 }
 
 // Advances to the table indicated by the index.  Returns false if there is
