+++ voltdbclient/serialization_helper.go
@@ -110,7 +110,7 @@ func serializeStatement(proc string, ud int64, args []driver.Value) (msg bytes.B
 			}
 			if _, ok := r.(error); ok {
 				err = r.(error)
-			}else {
+			} else {
 				fmt.Println(r)
 			}
